Refaktorerar dashboard.py, rensar onödig kod och flyttar load_css till utils (#58)
