+++ frontend/dashboard_pages/home_page.py
@@ -2,4 +2,11 @@ import streamlit as st
 
 
 def home():
-    st.markdown("# Startsida")
+    st.markdown(
+        '<h1 class="title-text">✨ Welcome to our Travel Robot Dashboard ✨</h1>',
+        unsafe_allow_html=True,
+    )
+    st.markdown(
+        '<h2 class="subheader-text">Your complete travel guide 🚆 🚋 🚌</h2>',
+        unsafe_allow_html=True,
+    )
